3.0.0 - Free version Changed to use one yml file per shop. If there is an old Shop.yml, it will be converted automatically. Improved to sort item indexes in the store yml file.
Moved 'Hide inventory' and 'Hide price type' to flags.
Added 'Hide Store Account Balance' flag.

Changed store location information to be automatically set when the localShop flag is set.

Fixed the problem that signShop only has coordinate values without world distinction.
Improved feedback when signShop setup fails.
When the shop UI is opened with a sign, the close button is changed to close the window unconditionally.

Changed so that a separate message is not sent when making a transaction that includes delivery charges.

Add Layout.yml.
Added some missing entries in the Lang file.
If there is a lore in the item meta, it is displayed in the shop UI.
Changed the flag in the shop information to be displayed only to admins.

Removed Convert command.

Tidy up the config file.
Other bug fixes and performance improvements.