Close button and page button can be changed. Various minor fixes and improvements. The icon of the close button and page button can be changed in config.yml.
Changed to cancel infinite stock when the median value of an item with unlimited stock is raised in the item edit window.

Fixed the problem that when dividing a value of -1 in the item edit window, it becomes 1.

Fixed a problem in which the sale price of an item was displayed slightly differently on the shop screen and the transaction screen. This happens with very low-stock items. It was just an error in the value displayed on the UI. (The number displayed on the transaction screen was the correct number.)
If you are curious about the reasons for this, the details are at the bottom.

Code refactoring and minor fixes

----------------------

Due to the nature of the way the plugin calculates the item price, even if the selling value and the purchase value are the same, the selling price and the purchase price at the same time are not perfectly equal.
Imagine a case where both prices are exactly the same. Here is an item that has a purchase price of $10 and a sale price of $10. When you buy one of these items, the price goes up because the number of items in the store is reduced. Now, if you resell the item you just bought, it's a very, Very small, but profitable. You can make money at no cost by buying items that are quite expensive in bulk and selling them immediately.