Inventory click search, etc. Now, clicking an item in your inventory on the start page will take you to the store that handles that item. (Left-click to move to the shop where you can purchase the item at the lowest price, and right-click to move to the shop where you can sell the item at the highest price.)
You can now click on inventory in the shop screen to go to the page with that item.

UI update related code optimization. The transaction screen is updated once every 5 seconds instead of once every 25 seconds. The shop UI now refreshes once every 5 seconds.

Change to return to the original page when you press close in the palette or trading UI

Fixed an issue that could sometimes fail to remove items from the store

Improved logic for Quick Sell to find the most advantageous stores for sellers

Added option to hide shops that users do not have access to in command auto-completion