3.3.3 You can now disable the shop. It is recommended to disable it first when editing the shop. If the shop data is changed while users are using it, related UI will be safely closed with a notification message. Newly created shops are created in an disabled state. Existing shops remain active.
Add 'hidden in command' flag

Fixed an issue that could sometimes fail to detect the economy plugin when the server is started.

Fixed a problem that occurred when the stock exceeds about 2.1 billion by selling items in the shop.

Fixed the price of items with a very low value and a very large quantity to not be zero.